docs: 项目全面分析完成，更新进度文档
- 使用8个并行智能体进行全面代码审查
- 更新项目进度约80%
- 新增游戏化系统（挑战+抽卡）
- 新增情绪分析、天气适配、语音功能框架
- 完善心率监测模块（数据仓库+配置页面）
- 完善GPS追踪和数据可视化

各模块完成度：
- 架构基础设施: 95%
- 记事本模块: 75%
- 提醒模块: 50% (缺导航入口)
- 运动模块: 95%
- 计划模块: 70%
- AI教练模块: 95%
- 心率监测: 90%
- 游戏化系统: 95%

+++ lib/features/workout/presentation/widgets/workout_bar_chart.dart
@@ -79,7 +79,7 @@ class WorkoutBarChart extends StatelessWidget {
             child: BarChart(
               BarChartData(
                 alignment: BarChartAlignment.spaceAround,
-                maxY: yMax,
+                maxY: yMax.toDouble(),
                 minY: 0,
                 groupsSpace: 12,
                 barTouchData: BarTouchData(
@@ -177,7 +177,7 @@ class WorkoutBarChart extends StatelessWidget {
                   },
                 ),
                 borderData: FlBorderData(show: false),
-                barGroups: _buildBarGroups(yMax),
+                barGroups: _buildBarGroups(yMax.toDouble()),
               ),
             ),
           ),
@@ -358,9 +358,9 @@ class WorkoutTypeBarChart extends StatelessWidget {
             child: BarChart(
               BarChartData(
                 alignment: BarChartAlignment.spaceAround,
-                maxY: yMax,
+                maxY: yMax.toDouble(),
                 minY: 0,
-                barGroups: _buildHorizontalBarGroups(yMax),
+                barGroups: _buildHorizontalBarGroups(yMax.toDouble()),
                 titlesData: FlTitlesData(
                   show: true,
                   leftTitles: AxisTitles(
